fix path aliases and move to app router
