--- apps/web/pages/_app.tsx
@@ -1,11 +1,0 @@
-import type { AppProps } from 'next/app';
-import { ClerkProvider } from '@clerk/nextjs';
-import '../styles/globals.css';
-
-export default function MyApp({ Component, pageProps }: AppProps) {
-  return (
-    <ClerkProvider {...pageProps}>
-      <Component {...pageProps} />
-    </ClerkProvider>
-  );
-}
