+++ apps/web/src/app/sign-up/[[...sign-up]]/page.tsx
