+++ apps/web/src/app/super/page.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { useState } from 'react';
 import { SignedIn, SignedOut, SignInButton } from '@clerk/nextjs';
