+++ tsconfig.base.json
@@ -16,7 +16,6 @@
     "baseUrl": ".",
     "jsx": "preserve",
     "paths": {
-      "@utils": ["packages/utils"],
       "@/*": ["packages/*", "apps/web/*"]
     }
   },
